refactor: dashboardのリファクタ
