+++ next-app/src/app/dashboard/page.tsx
@@ -1,5 +1,8 @@
+"use server";
+
+import { AppHeader } from "@/components";
 import DashBoard from "./Dashboard";
 
-export default function Page() {
-  return <DashBoard />
+export default async function Page() {
+  return <DashBoard header={<AppHeader />} />
 }
