+++ next-app/src/auth/components/LoginForm.tsx
@@ -1,10 +1,11 @@
+"use server";
+
 import { signIn } from "@/auth"
 
-export function SigninForm() {
+export async function SigninForm() {
   return (
     <form
       action={async () => {
-        "use server"
         await signIn("github", { redirectTo: "/" })
       }}
     >
