+++ next-app/src/components/AppHeader.tsx
@@ -1,25 +1,25 @@
-"use client";
-
 import Link from "next/link";
 import { AppButton } from "./AppButton";
-import { User } from "@/schemas";
 import { projectsMock } from "@/__mock__/projects.Mock";
-import { useState } from "react";
+import { auth } from "@/auth";
 
 
-export function AppHeader({ user }: { user: User | null }) {
+export async function AppHeader() {
   /* original → /
-  const { session } = useAuth();
   const { projectsData } = useProject();
-
-  const { status } = session;
   const { projects } = projectsData;
   /* ← original */
 
   /* mock → */
-  const [status, ] = useState("active");
   const projects = projectsMock;
   /* ← mock */
+
+
+  const session = await auth();
+  const user = session?.user;
+
+  if (!user) return null;
+
 
   return (
     <header className="sticky top-0 inset-x-0 z-40 bg-white/80 backdrop-blur-md border-b border-slate-200">
@@ -30,14 +30,8 @@ export function AppHeader({ user }: { user: User | null }) {
               href="/user"
               className="group flex items-center gap-2 text-slate-600 hover:text-blue-600 transition-colors"
             >
-              <div className={`w-2 h-2 rounded-full ${status === "active" ? "bg-emerald-500" : "bg-slate-300"}`} />
-              <span className="group-hover:underline underline-offset-4">
-                {user === null
-                  ? "ユーザー情報が不明です"
-                  : status === "active" ? `${user.email}` : "ログイン"
-                }
-              </span>
-              {status === "active" && <span className="text-[10px] bg-emerald-50 text-emerald-700 px-1.5 py-0.5 rounded font-bold uppercase">Online</span>}
+              <div className="w-2 h-2 rounded-full bg-emerald-500" />
+              <span className="text-[10px] bg-emerald-50 text-emerald-700 px-1.5 py-0.5 rounded font-bold">Signed In</span>
             </Link>
           </li>
           <li>
